rename RepositoryListViewer.java to RepositoryListViewerService.java and change package it from src.main.java.com.mizio.repository to src.main.java.com.mizio.service
