+++ src/main/java/com/mizio/controller/AddQuestionController.java
@@ -10,7 +10,7 @@ import com.mizio.manager.ViewManager;
 import com.mizio.model.*;
 import com.mizio.pattern.PathPattern;
 import com.mizio.pattern.TitlePattern;
-import com.mizio.repository.RepositoryListViewer;
+import com.mizio.service.RepositoryListViewerService;
 import com.mizio.service.RepositoryService;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -26,7 +26,7 @@ import java.util.ResourceBundle;
 
 public class AddQuestionController implements Initializable {
 
-    private RepositoryListViewer repositoryListViewer = new RepositoryListViewer();
+    private RepositoryListViewerService repositoryListViewer = new RepositoryListViewerService();
 
     private RepositoryService repositoryService = new RepositoryService();
