+++ src/main/java/com/mizio/controller/AddSubjectController.java
@@ -8,7 +8,7 @@ import com.mizio.manager.ViewManager;
 import com.mizio.model.Subject;
 import com.mizio.pattern.PathPattern;
 import com.mizio.pattern.TitlePattern;
-import com.mizio.repository.RepositoryListViewer;
+import com.mizio.service.RepositoryListViewerService;
 import com.mizio.service.RepositoryService;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -25,7 +25,7 @@ public class AddSubjectController implements Initializable {
 
     private RepositoryService repositoryService = new RepositoryService();
 
-    private RepositoryListViewer repositoryListViewer = new RepositoryListViewer();
+    private RepositoryListViewerService repositoryListViewer = new RepositoryListViewerService();
 
     private EditSubjectController editSubjectController = new EditSubjectController();
