+++ src/main/java/com/mizio/controller/EditSubjectController.java
@@ -4,7 +4,7 @@ import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXTextField;
 import com.mizio.concurrency.TableViewSubjectThread;
 import com.mizio.model.Subject;
-import com.mizio.repository.RepositoryListViewer;
+import com.mizio.service.RepositoryListViewerService;
 import com.mizio.service.RepositoryService;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -18,7 +18,7 @@ import java.util.ResourceBundle;
 public class EditSubjectController implements Initializable {
 
     RepositoryService repositoryService = new RepositoryService();
-    RepositoryListViewer repositoryListViewer = new RepositoryListViewer();
+    RepositoryListViewerService repositoryListViewer = new RepositoryListViewerService();
     private static int subjectID;
     private static Subject subject;
     private static TableViewSubjectThread thread;
