+++ src/main/java/com/mizio/controller/EditTestController.java
@@ -6,7 +6,7 @@ import com.jfoenix.controls.JFXTextField;
 import com.mizio.concurrency.TableViewTestThread;
 import com.mizio.model.Subject;
 import com.mizio.model.Test;
-import com.mizio.repository.RepositoryListViewer;
+import com.mizio.service.RepositoryListViewerService;
 import com.mizio.service.RepositoryService;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -20,7 +20,7 @@ import java.util.ResourceBundle;
 public class EditTestController implements Initializable {
 
     RepositoryService repositoryService = new RepositoryService();
-    RepositoryListViewer repositoryListViewer = new RepositoryListViewer();
+    RepositoryListViewerService repositoryListViewer = new RepositoryListViewerService();
     private static Test test;
     private static TableViewTestThread thread;
